finance_backend/requirements.txt
Django
djangorestframework
psycopg2-binary
django-filter
djangorestframework-simplejwt
drf-yasg
python-dotenv